+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Modification du lien permettant de récupérer les PES Acquit (en bas du cycle de vie)
 - Ajout de la constante ACTES_TYPE_PJ_IS_MANDATORY permettant de rendre obligatoire le typage des enveloppes #313
 - Vérification du code du type des PJ vis à vis de la liste récupéré dans la classification #419
+- Le script cron/jour.php est remplacé par le script script/actes/actes-send-classification-for-all-authorities.php #308
 
 ### Corrections